requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic-settings>=2.2
python-multipart>=0.0.9
numpy>=1.26
joblib>=1.3
scikit-learn>=1.4
tqdm>=4.66
timm>=0.9.12
torch>=2.2
torchvision>=0.17
ftfy>=6.2
regex>=2023.12.25
pillow>=10.3
ultralytics>=8.2
clip @ git+https://github.com/openai/CLIP.git
